Updating Carousel and Project Display to work better on mobile
